Some fixes to make tests work correctly
