+++ backend/package.json
@@ -9,6 +9,7 @@
     "scripts": {
         "test": "jest",
         "test:ci": "jest --ci",
+        "test:coverage": "jest --coverage",
         "lint": "eslint .",
         "lint:ci": "eslint . --max-warnings 0",
         "build": "tsc --project tsconfig.production.json",
@@ -105,7 +106,6 @@
             "/node_modules/",
             "/build/"
         ],
-        "collectCoverage": true,
         "collectCoverageFrom": [
             "src/**/*.ts"
         ],
