+++ backend/src/utils/oAuthClient.test.ts
@@ -1,4 +1,4 @@
-describe('googleController', () => {
+describe('oAuthClient', () => {
     describe('getOAuthClient', () => {
         test.todo('Should generate new OAuth2Client');
         test.todo('Should set access token in to the client when given');
